Separate the classes TitleLabel and  SubTitleLabel from FirstViewController
