+++ PhotoFrame/PhotoFrame/Label/CustomStyle.swift
@@ -1,0 +1,13 @@
+//
+//  CustomStyle.swift
+//  PhotoFrame
+//
+//  Created by 윤지영 on 30/11/2018.
+//  Copyright © 2018 윤지영. All rights reserved.
+//
+
+import Foundation
+
+protocol CustomStyle {
+    func setDefaultStyle()
+}
